Just consider Product cards as User Stories in they contain a certain label.
Right now hardcoded in the config the labels are: US and Issue.
I have also deleted the consolidated TwoBoardsClient and TechBoard in client module. I have made client
module to know twoboards stuff so I make it dependant of config. It wasn't very useful to have it agnostic of
twoboards stuff.